cred ca e bun v0.2

+++ EVR_VCU_K144/include/thermistor_mux.h
@@ -17,7 +17,7 @@ extern "C"{
 ==================================================================================================*/
 
 typedef struct Thermistors{
-	uint32 ThermistorValues[THERMISTOR_BANKS][THERMISTORS_PER_BANK];
+	uint16 ThermistorValues[THERMISTOR_BANKS][THERMISTORS_PER_BANK]; // Warning la functile Adc pentru setup si read daca folosesc uint32
 	Dio_ChannelType BankSelectPins[THERMISTOR_BANKS];
 	uint16 BankSelectPinsID[THERMISTOR_BANKS];
 	Adc_GroupType BankReadChannels[THERMISTORS_PER_BANK];
